remove a group entirely from display when all tasks in the group are deleted

+++ src/components/TodoList.js
@@ -5,23 +5,22 @@ import { List } from "semantic-ui-react";
 class TodoList extends React.Component {
   render() {
     const allTasks = this.props.tasks;
-    const listGroups = Object.keys(allTasks).map(key => {
-      return (
-        <TodoListGroup
-          tasks={allTasks[key]}
-          group={key}
-          checkItem={this.props.checkItem}
-          deleteItem={this.props.deleteItem}
-        />
-      );
+    const listGroups = Object.keys(allTasks).map(group => {
+      if (allTasks[group].length > 0) {
+        return (
+          <TodoListGroup
+            tasks={allTasks[group]}
+            group={group}
+            checkItem={this.props.checkItem}
+            deleteItem={this.props.deleteItem}
+            key={group}
+          />
+        );
+      }
+      return false;
     });
     return <List as="ol">{listGroups}</List>;
   }
 }
 
 export default TodoList;
-
-// this.state.tasks = {
-//   react: [{id, task}, {id, task}]
-
-// }
